+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Spoon trait. This allows any part to function like a spoon upgrade. It will only work on pickaxes, same as before.
 ### Changed
 - (API) - Stats are now a Forge registry and should be registered using the Register event. Stat names with the `silentgear` namespace can continue to omit the namespace in JSON files, others should include it.
 ### Fixed